2025-04-24 Fixed all mixins transform Overwrite to Inject

+++ src/main/java/net/karen/mccourse/mixin/EnchantmentHelperMixin.java
@@ -12,6 +12,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class EnchantmentHelperMixin {
     @Inject(at = @At("HEAD"), method = "Lnet/minecraft/world/item/enchantment/EnchantmentHelper;getEnchantmentLevel(Lnet/minecraft/nbt/CompoundTag;)I", cancellable = true)
     private static void mccourse$getEnchantmentLevel(CompoundTag nbt, CallbackInfoReturnable<Integer> cir) {
-        cir.setReturnValue(Mth.clamp(nbt.getInt("lvl"), 0, Integer.MAX_VALUE));
+        cir.setReturnValue(Mth.clamp(nbt.getInt("lvl"), 0, Short.MAX_VALUE));
     }
 }
